Cause non-zero return code on java exception
Using the CLI of qualimap, the script returns a zero code even if an
error is thrown by tha java code.

This is undesirable for several reasons, for instance when using qualimap
as part of an automated pipeline.

The change introduced in 3adb716 does not go far enough, as the command
calling the java code doesn't return a non-zero value.

This commit should mitigate this.